Game Manager Added
A game manager was added, who manages the activation state of the target objects and oversees correct seqential action. If correct target in a sequence has been touched, reward is also managed. If incorrect target is touched, no reward is given.

Minor changes were made in the Target Generation file, so that targets are now part of the  Agent.